+++ src/day8/re_frame_10x/fx/clipboard.cljs
@@ -1,12 +1,15 @@
-(ns day8.re-frame-10x.fx.clipboard)
+(ns day8.re-frame-10x.fx.clipboard
+  (:require
+   [day8.re-frame-10x.fx.window :refer [popout-window]]))
 
 (defn copy!
   [text]
-  (let [el (.createElement js/document "textarea")]
+  (let [doc (or (some-> @popout-window .-document) js/document)
+        el (.createElement doc "textarea")]
     (set! (.-value el) text)
     (set! (-> el .-style .-position) "absolute")
     (set! (-> el .-style .-left) "-9999px")
-    (.appendChild (.-body js/document) el)
+    (.appendChild (.-body doc) el)
     (.select el)
-    (.execCommand js/document "copy")
-    (.removeChild (.-body js/document) el)))
+    (.execCommand doc "copy")
+    (.removeChild (.-body doc) el)))